The ShennongTool Class

Stores the results generated by a computational tool or algorithm (e.g., UCell, GSVA, AUCell) applied to data in a Shennong object.

Slots

method

Character. The name of the tool or method used (e.g., "UCell").

version

Character. Version of the tool used.

result

A list or matrix storing the primary result.

params

A list of parameters used when running the tool.

timestamp

POSIXct. The time when the tool was run.

status

Character. Status of the run ("success", "error", etc.).

citation

Character. Optional reference or citation for the tool.

misc

A list of additional metadata.